[top-test] Move entropy_src_kat to library
Move entropy_src_kat test to a library so that the test can be reused.
